requirements.txt
Flask
flask-cors
google-generativeai
pymongo
dnspython